move `Activity` to `/system`
