+++ packages/status-react/src/system/activity/index.tsx
@@ -1,0 +1,1 @@
+export { Activity } from './activity'
